refactor: move win packaging to dedicated class
